refactor: resolve merge conflicts
