+++ src/main/kotlin/com/crafto/crafto_backend/mapper/CustomerIssueMapper.kt
@@ -3,7 +3,7 @@ package com.crafto.crafto_backend.mapper
 import com.crafto.crafto_backend.dto.CategoryResponse
 import com.crafto.crafto_backend.entity.CustomerIssue
 import com.crafto.crafto_backend.entity.CustomerIssueStatus
-import com.crafto.crafto_backend.request.CustomerIssueRequest
+import com.crafto.crafto_backend.dto.CustomerIssueRequest
 import com.crafto.crafto_backend.response.CraftsmanOfferResponse
 import com.crafto.crafto_backend.response.CustomerIssueDetailsResponse
 import com.crafto.crafto_backend.response.CustomerIssueResponse
