+++ src/main/kotlin/com/crafto/crafto_backend/mapper/CustomerMapper.kt
@@ -1,8 +1,8 @@
 package com.crafto.crafto_backend.mapper
 
 import com.crafto.crafto_backend.entity.Customer
-import com.crafto.crafto_backend.request.CustomerRequest
-import com.crafto.crafto_backend.response.CustomerResponse
+import com.crafto.crafto_backend.dto.CustomerRequest
+import com.crafto.crafto_backend.dto.CustomerResponse
 
 fun Customer.toResponse() = CustomerResponse(
     id = id ?: throw Exception("issue not fount"),
