+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=crafto-backend
-server.port=8085
+server.port=8086
 spring.data.mongodb.uri=${MONGODB_CONNECTION}
 spring.data.mongodb.database=crafto_db
 # Firebase Configuration
